refactor: 블록을 생성하는 책임을 BlockFactory로 이동
```
        위치가 블록을 만드는 방법을 알아야 할까요?!
`LocationOfMines`가 지뢰의 위치를 알고 있으니, `LocationOfMines`에게 지뢰가 있니? 라고 물어보고 응답에 따라 Field에서 `Land` 또는 `Mine`을 생성하는건 어떨까요?

_Originally posted by @ohgillwhan in https://github.com/next-step/kotlin-minesweeper/pull/202#discussion_r1056783704_

```
