--- src/main/kotlin/domain/Cell.kt
@@ -1,7 +1,0 @@
-package domain
-
-import domain.block.Block
-
-class Cell(
-    val block: Block
-)
