+++ src/test/kotlin/domain/AbstractCoordinateTest.kt
@@ -1,7 +1,5 @@
 package domain
 
-import domain.block.Land
-import domain.block.Mine
 import domain.coord.AbstractCoordinate
 import domain.coord.RelativeCoordinate
 import org.assertj.core.api.Assertions.assertThat
@@ -10,30 +8,6 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 
 internal class AbstractCoordinateTest {
-
-    @Test
-    fun `블록을 만드려는 곳에 지뢰가 존재하면 Mine이 반환횐다`() {
-        // given
-        val coordinate = AbstractCoordinate(1, 2)
-        val mines = LocationOfMines(listOf(AbstractCoordinate(1, 2)))
-        // when
-        val block = coordinate.toBlock(mines)
-
-        // then
-        assertThat(block).isInstanceOf(Mine::class.java)
-    }
-
-    @Test
-    fun `블록을 만드려는 곳에 지뢰가 없으면 Land 가 반환횐다`() {
-        // given
-        val coordinate = AbstractCoordinate(1, 2)
-        val mines = LocationOfMines(listOf(AbstractCoordinate(0, 0)))
-        // when
-        val block = coordinate.toBlock(mines)
-
-        // then
-        assertThat(block).isInstanceOf(Land::class.java)
-    }
 
     @ParameterizedTest
     @CsvSource(value = ["0, 0, -1, 1", "0, 0, 0, -1"])
